requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.0.6
django-extensions==3.2.3
django-taggit==5.0.1
executing==2.0.1
ipython==8.25.0
jedi==0.19.1
Markdown==3.6
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
